src/requirements.txt
# Requirements for Body Signals of Smoking project
# Tested on Python 3.10-3.11; use exact versions for model compatibility

streamlit==1.28.0
pandas==2.0.3
scikit-learn==1.4.1.post1
pillow==10.0.0
boto3==1.28.0
joblib==1.4.2